requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-multipart==0.0.9
numpy==1.26.4
scipy==1.13.1
torch==2.2.2
torchaudio==2.2.2
faster-whisper==1.0.3
pydantic==2.7.4
pydantic-settings==2.3.4
librosa==0.10.2
soundfile==0.12.1

# Mamba SSM (requires CUDA + nvcc — Linux with NVIDIA GPU only)
# Install manually after torch on a CUDA machine:
#   pip install causal-conv1d>=1.4.0 --no-build-isolation
#   pip install mamba-ssm --no-build-isolation
# On macOS, the GRU fallback in app/models/mamba_audio.py is used automatically.
